// src/supabase/types.ts
export type Database = {
// ...
	public: {
		Tables: {
			categories: {
				Row: {
					code: string
					description: string
					id: string
					name: string
				}
				Insert: {
					code: string
					description: string
					id?: string
					name: string
				}
				Update: {
					code?: string
					description?: string
					id?: string
					name?: string
				}
				Relationships: []
			}
			cities: {
				Row: {
					country: string
					cover_image: string
					description: string
					id: string
					location: unknown
					name: string
				}
				Insert: {
					country: string
					cover_image: string
					description: string
					id?: string
					location?: unknown
					name: string
				}
				Update: {
					country?: string
					cover_image?: string
					description?: string
					id?: string
					location?: unknown
					name?: string
				}
				Relationships: []
			}
			city_categories: {
				Row: {
					category_id: string
					city_id: string
				}
				Insert: {
					category_id: string
					city_id: string
				}
				Update: {
					category_id?: string
					city_id?: string
				}
				Relationships: [
					{
						foreignKeyName: 'city_categories_category_id_fkey'
						columns: ['category_id']
						isOneToOne: false
						referencedRelation: 'categories'
						referencedColumns: ['id']
					},
					{
						foreignKeyName: 'city_categories_category_id_fkey'
						columns: ['category_id']
						isOneToOne: false
						referencedRelation: 'cities_with_categories'
						referencedColumns: ['category_id']
					},
					{
						foreignKeyName: 'city_categories_city_id_fkey'
						columns: ['city_id']
						isOneToOne: false
						referencedRelation: 'cities'
						referencedColumns: ['id']
					},
					{
						foreignKeyName: 'city_categories_city_id_fkey'
						columns: ['city_id']
						isOneToOne: false
						referencedRelation: 'cities_with_categories'
						referencedColumns: ['id']
					}
				]
			}
			city_cities: {
				Row: {
					city_id: string
					related_city_id: string
				}
				Insert: {
					city_id: string
					related_city_id: string
				}
				Update: {
					city_id?: string
					related_city_id?: string
				}
				Relationships: [
					{
						foreignKeyName: 'city_cities_city_id_fkey'
						columns: ['city_id']
						isOneToOne: false
						referencedRelation: 'cities'
						referencedColumns: ['id']
					},
					{
						foreignKeyName: 'city_cities_city_id_fkey'
						columns: ['city_id']
						isOneToOne: false
						referencedRelation: 'cities_with_categories'
						referencedColumns: ['id']
					},
					{
						foreignKeyName: 'city_cities_related_city_id_fkey'
						columns: ['related_city_id']
						isOneToOne: false
						referencedRelation: 'cities'
						referencedColumns: ['id']
					},
					{
						foreignKeyName: 'city_cities_related_city_id_fkey'
						columns: ['related_city_id']
						isOneToOne: false
						referencedRelation: 'cities_with_categories'
						referencedColumns: ['id']
					}
				]
			}
			spatial_ref_sys: {
				Row: {
					auth_name: string | null
					auth_srid: number | null
					proj4text: string | null
					srid: number
					srtext: string | null
				}
				Insert: {
					auth_name?: string | null
					auth_srid?: number | null
					proj4text?: string | null
					srid: number
					srtext?: string | null
				}
				Update: {
					auth_name?: string | null
					auth_srid?: number | null
					proj4text?: string | null
					srid?: number
					srtext?: string | null
				}
				Relationships: []
			}
			tourist_attractions: {
				Row: {
					city_id: string | null
					description: string
					id: string
					name: string
				}
				Insert: {
					city_id?: string | null
					description: string
					id?: string
					name: string
				}
				Update: {
					city_id?: string | null
					description?: string
					id?: string
					name?: string
				}
				Relationships: [
					{
						foreignKeyName: 'tourist_attractions_city_id_fkey'
						columns: ['city_id']
						isOneToOne: false
						referencedRelation: 'cities'
						referencedColumns: ['id']
					},
					{
						foreignKeyName: 'tourist_attractions_city_id_fkey'
						columns: ['city_id']
						isOneToOne: false
						referencedRelation: 'cities_with_categories'
						referencedColumns: ['id']
					}
				]
			}
		}
		Views: {
			cities_with_categories: {
				Row: {
					category_code: string | null
					category_description: string | null
					category_id: string | null
					category_name: string | null
					country: string | null
					cover_image: string | null
					description: string | null
					id: string | null
					location: unknown
					name: string | null
				}
				Relationships: []
			}
			geography_columns: {
				Row: {
					coord_dimension: number | null
					f_geography_column: unknown
					f_table_catalog: unknown
					f_table_name: unknown
					f_table_schema: unknown
					srid: number | null
					type: string | null
				}
				Relationships: []
			}
			geometry_columns: {
				Row: {
					coord_dimension: number | null
					f_geometry_column: unknown
					f_table_catalog: string | null
					f_table_name: unknown
					f_table_schema: unknown
					srid: number | null
					type: string | null
				}
				Insert: {
					coord_dimension?: number | null
					f_geometry_column?: unknown
					f_table_catalog?: string | null
					f_table_name?: unknown
					f_table_schema?: unknown
					srid?: number | null
					type?: string | null
				}
				Update: {
					coord_dimension?: number | null
					f_geometry_column?: unknown
					f_table_catalog?: string | null
					f_table_name?: unknown
					f_table_schema?: unknown
					srid?: number | null
					type?: string | null
				}
				Relationships: []
			}
		}
// ...
	}
}